#30: Correctly identify and handle failed LilyPond compilations
Add function is_compiled

This is first used to determine whether a score has to be recompiled.
Look for either OUTPUT.pdf (for fullpage) or a non-empty
OUTPUT-systems.tex (note that also failed LilyPond compilations
produce a -systems.tex file)

the -systems.tex file also has to be removed, otherwise next time it would
be considered as "already compiled".