Create stub for a manual
It is important that - now that we start to have really many options -
there is proper documentation, and that we try to keep that up with
any new feature.
I think this has to be a manual and can't simply be the README.md anymore.

This creates a stub where any further documentation may be added to.

NOTE: For some reason I don't understand the main font gets lost somewhere
in the Pandoc chain. I have no idea what's wrong but we should fix that
as score files we may include will look awful that way.