Add 'twoside' option
Closes #119

When the 'twoside' option is not set (or reset to 'default') full-page scores will inherit the one/twoside option from the .tex document.
When set to either false or true that value is used regardles of the text
document's setting.

One special thing had to be considered is LilyPond's behaviour to respect
'inner-margin' only in two-side, 'left-margin' only in one-side mode.
The solution was to specify both as equal.
It *may* be desirable to clean that up a bit and write the full code depending
on the twoside option instead of only replacing the 't' or 'f'.