Alignment of inline scores (add 'valign' option)
The vertical alignment of inline (bare-inline) scores can now be
controlled with a combination of 'valign' and 'voffset'.

'valign' expects one out of center|top|bottom, while 'voffset' can be
any length.

The alignment is calculated from the bounding box of the system, but
so far (and probably forever) there's no way to determine antyhing
about the *content* of the image, specifically where the stafflines are.

Additionally a check is performed if the score has exactly one system,
and if it has more than one system a warning will be issued (normally
that doesn't make any sense).

This commit includes
Score:bbox(system) that can be used to retrieve the bounding box of
*any* system when compiled by systems,
Score:bbox(nil) will return the bounding box of the complete example (all
systems on top of each other).
This is mostly unused, only the bbox of the first system is used for the
vertical alignemnt, but maybe we can one day add more functionality
based on the height of the different systems.