Remove abstract
When thinking about what to write in the abstract I realized
that there's no substantial difference between the abstract and
the introduction (i.e. the introduction is actually as short as an
abstract).
But the introduction is just slightly too long for being an abstract,
so I thought I'd remove the abstract.