Implement max-protrusion
max-left-protrusion and max-right-protrusion work like "fences" that only
allow a certain amount of protrusion but do not interfere up to that point.
If the left or right protrusion is less than the max values they are taken
as-is, only after that point the score is compressed horizontally.

First a left offset is calculated, then it is determined if the score needs
recompilation (which generally is the case with justified scores, only
shorter fragments may simply be pushed to the right.
If necessary the score is immediately recompiled.

Negative max protrusion is possible, with an effect that is *similar*
to using quote and exampleindent/gutter. But while quote creates
hard limits the protrusion limits are dynamic. This is best seen by
consecutive scores with the same quote and (negative) protrusion
settings but different content (e.g. with and without instrument names).

The default value is currently set to 1cm, but I want to achieve a default
of the current margin, i.e. the page border acting as max protrusion "fence".