Delete remainders of failed LilyPond compilation
When a systems-by-systems compilation does *not* produce a score
it still leaves files around that are properly purged with this commit.

NOTE: I don't know if the os.execute call works on WIndows!

Also only scores that actually compiled are written FILELIST
(Note: the check is if they produced a score, not if LilyPond reports
them as compiled successfully (there are many ways LilyPond
"fails" but still produces output).