Make autoindent default to true
autoindent is a very useful option that should be on by default.
It *does* use time for recompilation, *but*: this will only be applied
when a max-protrusion limit is set and reached, i.e. in circumstances
where the user has explicitly accepted recompilations.

Handing the noautoindent option seemed to be an inappropriate
complication, and I honestly think it isn't necessary for that option.