Add functionality to append examples to manual
This includes a first document from the examples directory to be
appended to the manual.
The idea is to condense the manual itself to a reference and have
the demonstration in sections of arbitrary length at the end.

NOTE: dynamic-indent.tex has to be compiled, otherwise it will cause
a LaTeX error.
For this to be properly handled an appropriate addition to the Makefile
has to be done.