Page numbering for appended examples
the example documents appended to the manual will now be
paginated by the main document.

In order to prevent duplicate page numbers the next commit will
create a basic class file for the examples that just ensures
- that they are not paginated
- that they have the same page layout as the manual
- that a few basic macros are available

Apart from that it should be as bare as possible. For example they
won't include the font settings since we generally want to have the
examples as generic as possible, e.g. use fontspec only if this is
related to the topic to be demonstrated.