Add first-page-number and print-first-page-number options
a)
There may be the need to force a first page number (i.e. the start
with an odd or even page), especially in combination with the raw-pdf
option, when the resulting score is not immediately included in a score.
I came across this in a case where I want to first generate a number of
scores and afterwards include them in a row of documents (=> a set
of all parts for a score). I know that the scores will all start on an odd page
(because I'll take care of that) so I want to force LilyPond to think so,
regardless of what lyluatex thinks the "current page" is.

b)
Incidentally I think there may be reasons where the first page number may
*not* be wanted, even when included in a book.