Gracefully handle failed LilyPond start with 'showfailed' option
I think the 'showfailed' option should not only cover cases when LilyPond
fails to produce a score but also when it can't be started.

With the showfailed option a failure to start LilyPond does not result
in an err() but in a warn() with the same text, and a box is written
into the document like the one reporting a failed score compilation.

With this option it is for example possible to have people successfully
compile a document without a proper LilyPond installation (useful when
they only care for the content or for technical assistance).