Store OPTIONS in lyluatex.lua
It is necessary to store the OPTIONS as a local variable in lyluatex.lua
and pass a reference to lyluatex-lib.lua whenever options are needed there.
This is necessary to be able to use the -lib together with multiple packages.
Each package must keep its own copy of the OPTIONS.