Factor out lyluatex-options.lua
I think it is a good idea to factor this part out into a separate module too,
because that makes it possible to use the lyluatex-lib module from other code
independently, and to have a dedicated module dealing with the option handling.

What will be needed is documentation for package authors on how to use the
-options module. Actually I think it would even be a good idea to provide that
as a separate Lua(LaTeX) package.