Store both declarations and options in lyluatex-options
Instead of storing the option declarations in lyluatex-options.lua and
the actual entries in lyluatex.lua there is now one OPTIONS object in
lyluatex-options.lua.
The top-level entries are the prefixes of client libraries ('ly' for
lyluatex), with 'declarations' and 'options' fields for each client.

For lyluatex this requires a somewhat awkward workaround to handle
the double usage of `Score` as both a data storage table and a metatable.
However, third-party clients of lyluatex-options can simply access both
declarations and options through optlib's functions now.