\setluaoption, automatically store registered clients
This is a generic version of \lysetoption and Opts:set_option()
addressing an arbitrary registered client.

Global variables <client-name>_opts are now stored automatically,
it's not the responsibility of the user.

TODO: Discuss whether these Opts objects should really be available
as global variables. Maybe it would be better to store them in a table
and provide an accessor function (although that needs slightly more complicated
code and might require updates to existing packages).