optlib.register()
Differentiate between
  - optlib.register()
and
  - Opts:new()

Opts:new() creates an Opts object and populates it with the
declarations.
optlib.register() is responsible for creating the global variable
and for handling the *package* options.

The idea is that it is necessary to allow an Opts object to be created for
option validation without doing it for a whole *package*.