Store registered clients locally
In order not to pollute the global Lua namespace any registered
client (Opts object) is now stored in a local table and can be
accessed through lua_options.client(name).
This will raise an error if not client with that name is registered.

Note: Existing clients can either change all their <client>_opts variables
or load them into a local variable (which is more efficient anyway).