"Manual"
For the initial release I think it's OK to go with a short introduction.
This is a "helper" package, after all. Of course it is necessary to have
proper "developer documentation" for the package to become useful, but
I think we can aim at a 1.0f release along with luaformatters v1.0